+++ apps/backoffice-v2/src/pages/Entity/components/Case/Case.Documents.tsx
@@ -7,7 +7,7 @@ import { ImageViewer } from '@/common/components/organisms/ImageViewer/ImageView
 import { ctw } from '@/common/utils/ctw/ctw';
 import { keyFactory } from '@/common/utils/key-factory/key-factory';
 import { DocumentsToolbar } from '@/pages/Entity/components/Case/Case.Documents.Toolbar';
-import { useDocuments } from './hooks/useDocuments/useDocuments';
+import { useDocumentsLogic } from './hooks/useDocuments/useDocumentsLogic';
 import { IDocumentsProps } from './interfaces';
 
 /**
@@ -24,7 +24,10 @@ import { IDocumentsProps } from './interfaces';
  */
 export const Documents: FunctionComponent<IDocumentsProps> = ({
   documents,
+  onOcrPressed,
   isLoading,
+  isDocumentEditable,
+  isLoadingOCR,
   hideOpenExternalButton,
 }) => {
   const {
@@ -32,11 +35,9 @@ export const Documents: FunctionComponent<IDocumentsProps> = ({
     onCrop,
     onCancelCrop,
     isCropping,
-    onOcr,
     selectedImageRef,
     initialImage,
     skeletons,
-    isLoadingOCR,
     selectedImage,
     onSelectImage,
     documentRotation,
@@ -45,8 +46,9 @@ export const Documents: FunctionComponent<IDocumentsProps> = ({
     onTransformed,
     isRotatedOrTransformed,
     shouldDownload,
+    isOCREnabled,
     fileToDownloadBase64,
-  } = useDocuments(documents);
+  } = useDocumentsLogic(documents);
 
   return (
     <ImageViewer selectedImage={selectedImage} onSelectImage={onSelectImage}>
@@ -88,8 +90,10 @@ export const Documents: FunctionComponent<IDocumentsProps> = ({
             onOpenDocumentInNewTab={onOpenDocumentInNewTab}
             // isRotatedOrTransformed={isRotatedOrTransformed}
             shouldDownload={shouldDownload}
+            isOCREnabled={!!isDocumentEditable && isOCREnabled}
+            onOcrPressed={onOcrPressed}
+            isLoadingOCR={!!isLoadingOCR}
             // isCropping={isCropping}
-            // isLoadingOCR={isLoadingOCR}
             // onCancelCrop={onCancelCrop}
             fileToDownloadBase64={fileToDownloadBase64}
           />